Add test ShouldGenerateInvoiceDiscounting6Percent

+++ Common/InvoiceCalculators/Order.cs
@@ -3,10 +3,10 @@ namespace Common.InvoiceCalculators
   public class Order
   {
     public string Customer { get; private set; }
-    public double TotalValue { get; private set; }
+    public decimal TotalValue { get; private set; }
     public int AmountItems { get; private set; }
 
-    public Order(string customer, double totalValue, int amountItems)
+    public Order(string customer, decimal totalValue, int amountItems)
     {
       Customer = customer;
       TotalValue = totalValue;
